Fix bug where action failures are swallowed with a nil inside an ensure.

+++ lib/kitchen/chef_data_uploader.rb
@@ -72,7 +72,7 @@ module Kitchen
       cookbooks_dir = local_cookbooks
       upload_path(scp, cookbooks_dir, "cookbooks")
     ensure
-      FileUtils.rmtree(cookbooks_dir)
+      FileUtils.rmtree(cookbooks_dir) unless cookbooks_dir.nil?
     end
 
     def upload_data_bags(scp)
